After making css changes for Candidate profile
